requirements.txt
google-genai
pyTelegramBotAPI
gTTS
pillow
python-dotenv
